zip file existence check has been added.
git-svn-id: https://tmp.openrtm.org/svn/rtmtools/trunk@96 16521fb8-224f-4d59-aec7-8b1ded20d512

+++ buildall.sh
@@ -80,7 +80,9 @@ for project in $projects; do
     fi
 done
 
-rm $JARDIR.zip
+if test -f $JARDIR.zip ; then
+    rm $JARDIR.zip
+fi
 zip $JARDIR.zip -r ./$JARDIR
 
 export ANT_HOME=$DUMMY
